test: a user can not delete himself

+++ app/Http/Controllers/Admin/UserController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Auth;
 
 class UserController extends Controller
 {
@@ -110,6 +111,19 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
+
+        //No permitir que el usuario logueado se borre a si mismo
+        if($user->id == Auth::user()->id){
+            session()->flash('swal', [
+                'icon' => 'error',
+                'title' => 'Accion Denegada',
+                'text' => 'No puedes eliminarte a ti mismo'
+            ]);
+            abort(403, 'No puedes borrar tu propio usuario');
+
+            return redirect(route('admin.users.index'));
+        }
+
         //Eliminar roles asociados
         $user->roles()->detach();
